# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "Dokumentasi API Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Ambil semua barang",
                "operationId": "11923c34cb676a35e28450679de6843e",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Tambah barang",
                "operationId": "9579148927fc19a376d89a9c8f88cbec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_barang",
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "kode_barang": {
                                        "type": "string",
                                        "example": "PH001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Lampu Philips"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Ambil barang berdasarkan ID",
                "operationId": "16764df0e9538934439e95cb7b02b78a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update seluruh data barang",
                "operationId": "f2b94eb93c0e8dafdf78958a89f9429a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_barang",
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "kode_barang": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil update"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Hapus barang",
                "operationId": "fcf7e384d2ed6ab2d8c5d56a39da057a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Items"
                ],
                "summary": "Update sebagian data barang",
                "operationId": "7884a4652d52925f1614d50fed40327d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_barang": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil patch"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Items",
            "description": "Items"
        }
    ]
}